Standardization, extension of visualizations, benchmark hub, and general updates

+++ tests/autotuning_methodology/integration/test_run_experiment.py
@@ -1,40 +1,36 @@
 """Integration test for running and fetching an experiment from cache."""
 
-import json
 from importlib.resources import files
 from pathlib import Path
 from shutil import copyfile
+from warnings import warn
 
 import numpy as np
 import pytest
-from jsonschema import validate
 
 from autotuning_methodology.curves import StochasticOptimizationAlgorithm
-from autotuning_methodology.experiments import (
-    ResultsDescription,
-    execute_experiment,
-    get_args_from_cli,
-    get_experiment_schema_filepath,
-)
+from autotuning_methodology.experiments import ResultsDescription, execute_experiment, get_args_from_cli
+from autotuning_methodology.validators import validate_experimentsfile
 
 # get the path to the package
 package_path = Path(files("autotuning_methodology")).parent.parent
-# package_path = ""
 
 # setup file paths
+strategy = "random_sample_10_iter"
 mockfiles_path_root = package_path / Path("tests/autotuning_methodology/integration/mockfiles/")
 mockfiles_path_source = mockfiles_path_root / "mock_gpu.json"
 mockfiles_path = mockfiles_path_root
 experiment_filepath_test = mockfiles_path / "test.json"
 assert experiment_filepath_test.exists()
 kernel_id = "mocktest_kernel_convolution"
-cached_visualization_path = package_path / Path(f"cached_data_used/visualizations/test_run_experiment/{kernel_id}")
-cached_visualization_file = cached_visualization_path / "mock_GPU_random_sample_10_iter.npz"
-cached_visualization_imported_path = package_path / Path(
-    f"cached_data_used/visualizations/test_output_file_writer/{kernel_id}"
-)
-cached_visualization_imported_file = cached_visualization_imported_path / "mock_GPU_ktt_profile_searcher.npz"
-normal_cachefiles_path = package_path / Path(f"cached_data_used/cachefiles/{kernel_id}")
+experiment_path = package_path / Path("test_run_experiment")
+experiment_path_run = experiment_path / "run"
+experiment_path_setup = experiment_path / "setup"
+
+cached_visualization_path = experiment_path_run
+plot_path = cached_visualization_path / "generated_graphs"
+cached_visualization_file = experiment_path_run / strategy / "mock_GPU_mocktest_kernel_convolution.npz"
+normal_cachefiles_path = package_path / Path(f"benchmark_hub/cachefiles/{kernel_id}")
 normal_cachefile_destination = normal_cachefiles_path / "mock_gpu.json"
 experiment_import_filepath_test = mockfiles_path / "test_import_runs.json"
 assert experiment_import_filepath_test.exists()
@@ -43,12 +39,24 @@ import_runs_path = package_path / Path("cached_data_used/import_runs")
 import_runs_filepaths: list[Path] = list()
 
 
-def _remove_dir(path: Path):
+def _remove_dir(path: Path, ignore_permission_error=False):
     """Utility function for removing a directory and the contained files."""
-    assert path.exists()
+    assert path.exists(), f"Path to directory does not exist: {path.resolve()}"
+    permission_errors = []
     for sub in path.iterdir():
-        sub.unlink()
-    path.rmdir()
+        try:
+            if sub.is_dir():
+                _remove_dir(sub)
+            else:
+                sub.unlink()
+        except PermissionError as e:
+            if ignore_permission_error:
+                warn(e)
+                permission_errors.append(e)
+            else:
+                raise e
+    if not (ignore_permission_error and len(permission_errors) > 0):
+        path.rmdir()
 
 
 def setup_module():
@@ -58,9 +66,6 @@ def setup_module():
     assert normal_cachefiles_path.exists()
     normal_cachefile_destination.write_text(mockfiles_path_source.read_text())
     assert normal_cachefile_destination.exists()
-    # cached_visualization_path.mkdir(parents=True, exist_ok=True)
-    # assert cached_visualization_path.exists()
-    # copy the import run test files to the import run folder
     assert import_runs_source_path.exists()
     import_runs_path.mkdir(parents=True, exist_ok=True)
     assert import_runs_path.exists()
@@ -79,15 +84,10 @@ def teardown_module():
     if normal_cachefile_destination.exists():
         normal_cachefile_destination.unlink()
     _remove_dir(normal_cachefiles_path)
-    if cached_visualization_file.exists():
-        cached_visualization_file.unlink()
-    _remove_dir(cached_visualization_path)
-    if cached_visualization_imported_file.exists():
-        cached_visualization_imported_file.unlink()
-    _remove_dir(cached_visualization_imported_path)
     # delete the import run test files from the import run folder
     for import_run_file in import_runs_filepaths:
         import_run_file.unlink()
+    _remove_dir(experiment_path)
 
 
 def test_CLI_input():
@@ -96,11 +96,11 @@ def test_CLI_input():
     with pytest.raises(SystemExit) as e:
         dummy_args = ["-dummy_arg=option"]
         get_args_from_cli(dummy_args)
-    assert e.type == SystemExit
+    assert e.type == SystemExit  # noqa: E721
     assert e.value.code == 2
 
     # improper input 2
-    with pytest.raises(ValueError, match="Invalid '-experiment' option"):
+    with pytest.raises(ValueError, match="Invalid '--experiment' option"):
         get_args_from_cli([""])
 
     # proper input
@@ -122,7 +122,7 @@ def test_bad_experiment():
 def test_run_experiment_bad_kernel_path():
     """Run an experiment with a bad kernel path."""
     experiment_filepath = str(mockfiles_path / "test_bad_kernel_path.json")
-    with pytest.raises(FileNotFoundError, match="No such path"):
+    with pytest.raises(FileNotFoundError, match="does not exist"):
         execute_experiment(experiment_filepath, profiling=False)
 
 
@@ -133,8 +133,10 @@ def test_run_experiment():
     if cached_visualization_file.exists():
         cached_visualization_file.unlink()
     assert not cached_visualization_file.exists()
-    (experiment, strategies, results_descriptions) = execute_experiment(str(experiment_filepath_test), profiling=False)
-    validate_experiment_results(experiment, strategies, results_descriptions)
+    (experiment, all_experimental_groups, searchspace_statistics, results_descriptions) = execute_experiment(
+        str(experiment_filepath_test), profiling=False
+    )
+    validate_experiment_results(experiment, all_experimental_groups, searchspace_statistics, results_descriptions)
 
 
 @pytest.mark.usefixtures("test_run_experiment")
@@ -144,29 +146,23 @@ def test_cached_experiment():
     assert normal_cachefile_destination.exists()
     assert cached_visualization_path.exists()
     assert cached_visualization_file.exists()
-    (experiment, strategies, results_descriptions) = execute_experiment(str(experiment_filepath_test), profiling=False)
-    validate_experiment_results(experiment, strategies, results_descriptions)
-
-
-def test_import_run_experiment():
-    """Import runs from an experiment."""
-    assert import_runs_path.exists()
-    (experiment, strategies, results_descriptions) = execute_experiment(
-        str(experiment_import_filepath_test), profiling=False
+    (experiment, all_experimental_groups, searchspace_statistics, results_descriptions) = execute_experiment(
+        str(experiment_filepath_test), profiling=False
     )
-    assert cached_visualization_imported_path.exists()
-    assert cached_visualization_imported_file.exists()
-    validate_experiment_results(experiment, strategies, results_descriptions)
+    validate_experiment_results(experiment, all_experimental_groups, searchspace_statistics, results_descriptions)
 
 
 @pytest.mark.usefixtures("test_run_experiment")
 def test_curve_instance():
     """Test a Curve instance."""
     # setup the test
-    (experiment, strategies, results_descriptions) = execute_experiment(str(experiment_filepath_test), profiling=False)
-    kernel_name = experiment["kernels"][0]
-    gpu_name = experiment["GPUs"][0]
-    strategy_name = strategies[0]["name"]
+    (experiment, all_experimental_groups, _, results_descriptions) = execute_experiment(
+        str(experiment_filepath_test), profiling=False
+    )
+    experimental_groups: dict = experiment["experimental_groups_defaults"]
+    kernel_name = experimental_groups["applications"][0]["name"]
+    gpu_name = experimental_groups["gpus"][0]
+    strategy_name = all_experimental_groups[0]["name"]
     results_description = results_descriptions[gpu_name][kernel_name][strategy_name]
     curve = StochasticOptimizationAlgorithm(results_description)
 
@@ -194,21 +190,31 @@ def test_curve_instance():
 
 def validate_experiment_results(
     experiment,
-    strategies,
+    all_experimental_groups,
+    searchspace_statistics,
     results_descriptions,
 ):
     """Validate the types and contents returned from an experiment."""
-    assert isinstance(experiment, dict)
-    assert isinstance(strategies, list)
-    assert isinstance(results_descriptions, dict)
+    assert isinstance(experiment, dict), f"should be dict, is {type(experiment)} ({experiment})"
+    assert isinstance(
+        searchspace_statistics, dict
+    ), f"should be dict, is {type(searchspace_statistics)} ({searchspace_statistics})"
+    assert isinstance(
+        all_experimental_groups, list
+    ), f"should be list, is {type(all_experimental_groups)} ({all_experimental_groups})"
+    assert isinstance(
+        results_descriptions, dict
+    ), f"should be dict, is {type(results_descriptions)} ({results_descriptions})"
 
     # validate the contents
-    schemafilepath = get_experiment_schema_filepath()
-    with open(schemafilepath, "r", encoding="utf-8") as schemafile:
-        schema = json.load(schemafile)
-        validate(instance=experiment, schema=schema)
-    kernel_name = experiment["kernels"][0]
-    gpu_name = experiment["GPUs"][0]
-    assert len(strategies) == 1
-    strategy_name = strategies[0]["name"]
+    validate_experimentsfile(experiment)
+    experimental_groups: dict = experiment["experimental_groups_defaults"]
+    assert isinstance(experimental_groups, dict)
+    kernel_name = experimental_groups["applications"][0]["name"]
+    assert kernel_name == "mocktest_kernel_convolution"
+    gpu_name = experimental_groups["gpus"][0]
+    assert gpu_name == "mock_GPU"
+    assert len(all_experimental_groups) == 1
+    strategy_name = all_experimental_groups[0]["name"]
+    assert strategy_name == "random_sample_10_iter"
     assert isinstance(results_descriptions[gpu_name][kernel_name][strategy_name], ResultsDescription)
